requirements.txt
python-dotenv
pandas
openai
sqlalchemy
streamlit
